requirements.txt
# Core dependencies
flask
python-dotenv
-e .

# LangChain ecosystem
langchain
langchain_community
langchain_pinecone
langchain_openai
langchain_experimental

# Vector DB
pinecone

# LLMs and Transformers
torch>=2.1.0
transformers>=4.38
accelerate

# Embeddings
sentence-transformers==2.2.2

# PDF support
pypdf
